destructure modal and signup form
