+++ public/index.html
@@ -6,7 +6,7 @@
   <title>Big Poppa Code Starter Kit</title>
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="stylesheet" href="/css/styles.css">
-<script defer src="/js/dist/vendors-node_modules_css-loader_dist_runtime_api_js-node_modules_css-loader_dist_runtime_sour-354ecd.5c2fb523829af820f1f1.js?56367248a4c615f86ac8"></script><script defer src="/js/dist/App.6f6f24b5c760056d1958.js?56367248a4c615f86ac8"></script><script defer src="/js/dist/main.e70a9980834bdea1725b.js?56367248a4c615f86ac8"></script></head>
+<script defer src="/js/dist/vendors-node_modules_css-loader_dist_runtime_api_js-node_modules_css-loader_dist_runtime_sour-354ecd.5c2fb523829af820f1f1.js?e10d7b33df677035f65e"></script><script defer src="/js/dist/App.29aab886a7fb367012b7.js?e10d7b33df677035f65e"></script><script defer src="/js/dist/main.e70a9980834bdea1725b.js?e10d7b33df677035f65e"></script></head>
 
 <body>
   <div id="app"></div>
