+++ src/components/AuthModal/AuthModal.module.scss
@@ -1,0 +1,12 @@
+.dialog {
+    width: 35%;
+    border: .1rem solid grey;
+    border-radius: .5rem;
+    box-shadow: 3px 3px 5px 1px rgba(93, 128, 80, 0.99);
+    overflow: hidden;
+}
+
+.dialog::backdrop {
+    backdrop-filter: blur(2px);
+    background-color: rgba(128, 120, 120, 0.3);
+}
