+++ src/components/signUpForm/signUpForm.module.scss
@@ -1,15 +1,3 @@
-.dialog {
-    width: 35%;
-    border: .1rem solid grey;
-    border-radius: .5rem;
-    box-shadow: 3px 3px 5px 1px rgba(93, 128, 80, 0.99);
-}
-
-.dialog::backdrop {
-    backdrop-filter: blur(3px);
-    background-color: rgba(33, 32, 32, 0.3);
-}
-
 .h1 {
     text-align: center;
     font-size: 3.5rem;
